Handle primitive return types
There was a bug in the code generation when the return type of an
operation was a primitive. The code generation would generate
something like Future<int>, which doesn't compile. The new code
generation uses a boxed value instead, e.g. Future<java.lang.Integer>.

Added unit tests for primitive types and parameters.